+++ src/pyobsplot/utils.py
@@ -7,6 +7,9 @@ import pathlib
 # Output directory of esbuild
 bundler_output_dir = pathlib.Path(__file__).parent / "static"
 
+# Minimum npm package version
+min_npm_version = "0.3.3"
+
 # Allowed default values
 allowed_defaults = [
     "marginTop",