+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - `jsdom` renderer now uses a local http server instead of calling a script at each invocation, greatly improving rendering speed.
 - Autocompletion of Plot methods should now be working in IDEs.
+- Ensure that the needed version of the npm package is run if jsdom renderer is used.
 - Debug mode also works with `jsdom` renderer.
 - Plot generator objects now have correct `__repr__` methods.
 - Update anywidget to 0.2.3.